requirements.txt
openai>=1.0.0
pandas
tqdm
scikit-learn
pyyaml
python-dotenv
